add_handle_tool_top_menu

+++ src/components/Smart/DrawWindow/RightPanelContent/Frames/Frames.jsx
@@ -27,7 +27,7 @@ function Frames(props) {
 
     function deleteSelectedFrame(){
         setMenuActive(false);
-        FrameManager.deleteSelectedFrame();
+        if(window.confirm('Удалить выбранные кадры ?')) FrameManager.deleteSelectedFrame();
     }
     let FrameMenu = <ContextMenu active={menuActive}
                                  setMenuActive={setMenuActive}
